backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
pdfplumber>=0.10.2
pdf2image>=1.17.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
transformers>=4.41.1
torch>=2.3.0
Pillow>=10.3.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
gunicorn>=21.2.0
whitenoise>=6.6.0

